Remove linux/target source file branching
